feat(#160): Migrate schemas to TypeScript with inferred types
Phase 1 Progress: schemas.mjs → schemas.ts (2/5 complete)

Changes:
- ✅ Convert schemas.mjs to schemas.ts
- ✅ Add type exports using z.infer<typeof Schema> for all schemas
- ✅ Export explicit type aliases for enums (TransactionType, Category, etc.)
- ✅ Maintain backward compatibility (BatchConfigSchema alias)
- ✅ Build output: schemas.mjs (4.2kb)

Type Safety Benefits:
- Type-safe schema definitions with Zod
- Inferred TypeScript types from schemas
- IDE autocomplete for Transaction, OcrResult, ModelResult, etc.
- Compile-time validation for schema usage

Schema Types Exported:
- OcrResult, ModelResult, Transaction
- AzureCredentials, SystemConfig
- TransactionType, TransactionStatus, Category, MerchantSource

Next: model-analyzer.mjs (456 lines, AI model comparison logic)
